DOCKER-CONTROLLER/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
docker==7.1.0
idna==3.11
pip==26.0.1
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.3